requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.4
httpx==0.26.0
pandas==2.1.4
